Perfusion phantom with multiple pools.
This phantom is a sort of variation on the previous design meant to replace the use of sponges with geometric shapes
that act as 'pools.' A bolus of contrast agent circulating through the main flow line should initially diffuse into the
pool, and then when the bolus passes, the pooled contrast agent should diffuse out. The diffusion rate should depend on
the geometry, contrast agent concetration / shape of bolus, and the flow rate.

However, this phantom, like it's predecessor, turned out to be too finicky to print with the available printers. The
multiple pools also have a termporal smearing effect since the upstream pools impact the diffusion of all later pools.