Added a housing for a 'halvklart' button lamp.
This housing is designed to be mounted above a desk, but confine (most) light to the kayboard area mostly underneath it.
It works well. Even when printed with translucent filament the light spray is tolerable.

Note that the 'v0' stl is slightly different compared with the source. I tweaked the source file without re-exporitng
the STL and forget to rename. I think only the deflector arm position was tweaked to slightly broaden the lit volume.