Added a mounting bracket.
This bracket encloses a light, so it must be printed by splitting in half and printing in two parts. Then the two halves
can be joined together after printing to enclose the light. The bracket is symmetric (hopefully!) so only one half needs
to be recovered, so a simple boolean operation will suffice.