Modifications after fit testing.
Fitted on a 2008 Specialized SX Trail 2.

Shrunk the plate size so it has lower risk of interfering with the chainstay/pivot body, added slots for the upper guide
so it can slide closer to the chainring, added a nother set of holes for the tensioner so it can move 1cm closer to the
chainring (and thus increase tension in the chain), and shrunk the tensioner so the RHS pedal can (hopefully!) clear it.